+++ src/app/doctor/disease/disease.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { ConfirmationService, MessageService } from 'primeng/api';
 import { DiseaseService } from 'src/app/services/disease.service';
 
 @Component({
@@ -10,7 +11,8 @@ export class DiseaseComponent implements OnInit {
 
   dtOptions: DataTables.Settings = {};
   diseaseList:{}
-  constructor(private diseaseService:DiseaseService) { }
+  constructor(private diseaseService:DiseaseService,private confirmationService: ConfirmationService,
+    private messageService: MessageService) { }
 
   ngOnInit(): void {
     this.dtOptions = {
@@ -24,9 +26,26 @@ export class DiseaseComponent implements OnInit {
     })
   }
   delete(value){
-    this.diseaseService.deleteDisease(value).subscribe(res =>{
-      console.log("Disease Deleted!!");
 
-    })
+    this.confirmationService.confirm({
+      message: 'Are you sure that you want to proceed?',
+      header: 'Confirmation',
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.diseaseService.deleteDisease(value).subscribe(res =>{
+          console.log("Disease Deleted!!");
+
+        })          },
+      reject: () => {
+        this.messageService.add({
+          severity: 'error',
+          summary: 'Rejected',
+          detail: 'You have rejected',
+        });
+      },
+    });
+
+
+
   }
 }
